+++ e2e/commands/status.e2e.2.ts
@@ -11,6 +11,7 @@ import {
   importPendingMsg
 } from '../../src/cli/commands/public-cmds/status-cmd';
 import * as fixtures from '../../src/fixtures/fixtures';
+import { MISSING_DEPS_SPACE, MISSING_NESTED_DEPS_SPACE } from '../../src/constants';
 
 const assertArrays = require('chai-arrays');
 
@@ -75,14 +76,34 @@ describe('bit status command', function() {
     let output;
     before(() => {
       helper.scopeHelper.reInitLocalScope();
-      helper.fixtures.createComponentBarFoo();
-      helper.fs.createFile('bar', 'foo2.js', 'var foo = require("./foo.js")');
-      helper.command.addComponent('bar/foo2.js', { i: 'bar/foo2' });
+      helper.fs.createFile(
+        '',
+        'comp1.js',
+        `require("./comp2");require("./comp3");require("./comp4");require("./comp5");require("./comp6");`
+      );
+      helper.fs.createFile('', 'comp2.js', `require("./comp4");require("./comp5");`);
+      helper.fs.createFile('', 'comp3.js', '');
+      helper.fs.createFile('', 'comp4.js', '');
+      helper.fs.createFile('', 'comp5.js', 'require("./comp6");');
+      helper.fs.createFile('', 'comp6.js', '');
+      helper.command.addComponent('comp1.js', { i: 'comp1' });
+      helper.command.addComponent('comp5.js', { i: 'comp5' });
     });
     it('Should show missing dependencies', () => {
       output = helper.command.runCmd('bit status');
+      const countComp5Deps = (output.match(/comp5.js -> comp6.js/g) || []).length;
       expect(output).to.have.string('untracked file dependencies');
-      expect(output).to.have.string('bar/foo2.js -> bar/foo.js');
+      expect(output).to.have.string('comp1 ...  issues found');
+      expect(output).to.have.string('comp1.js -> comp2.js, comp3.js, comp4.js, comp6.js');
+      expect(output).to.have.string('comp2.js -> comp4.js');
+      expect(output).to.have.string('comp5.js -> comp6.js');
+      // Make sure comp5->comp6 appear only under comp5
+      expect(countComp5Deps).to.equal(1);
+      expect(output).to.have.string('comp5 ...  issues found');
+      // Validate indentations is correct, nested deps should be indent 2 more
+      expect(output).to.have.string(`${MISSING_DEPS_SPACE}comp1.js`);
+      expect(output).to.have.string(`${MISSING_NESTED_DEPS_SPACE}comp2.js`);
+      expect(output).to.have.string(`${MISSING_DEPS_SPACE}comp5.js`);
     });
   });
   describe('when a component is created and added without its package dependencies', () => {