+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- make bit status recursively show untracked files dependencies
+
 ## [14.6.1-dev.1] - 2019-12-04
 
 - add isolate function to testers new API